refactor codegen
* split into separate files
* extract some methods
* fix deprecated string concatenations: use interpolation

preparation for moving this to overflowdb to make it reusable
