+++ project/meta-build.sbt
@@ -1,4 +1,4 @@
 libraryDependencies ++= Seq(
   "com.github.pathikrit" %% "better-files" % "3.8.0",
-  "com.typesafe.play" %% "play-json" % "2.6.8"
+  "com.typesafe.play" %% "play-json" % "2.8.1"
 )
